[RLlib] Fix test_ope flakiness (#27676)
